requirements.txt
psycopg2-binary==2.9.1
pip~=19.0.3
setuptools~=40.8.0
python-dotenv~=0.18.0
django==3.1
django-model-utils==4.0.0
psycopg2-binary==2.9.1
pip~=19.0.3
setuptools~=40.8.0
python-dotenv~=0.18.0
factory-boy==3.2.0
PyPrind==2.11.3
flake8==3.9.2
pytz~=2021.1
PyYAML~=5.4.1